build specify extension

+++ build.sh
@@ -1,5 +1,10 @@
 #!/usr/bin/sh
 
+if [ "$1" != "chrome" ] && [ "$1" != "firefox" ]; then
+  echo "Usage: ./build.sh chrome|firefox"
+  exit 1
+fi
+
 rm build.zip
 mkdir -p build
 cp manifest.json build/
@@ -7,6 +12,11 @@ cp -r images build/
 cp -r scripts build/
 cp -r styles build/
 cd build || exit
+
+if [ "$1" = "chrome" ]; then
+  sed -i 's/"manifest_version": 2,/"manifest_version": 3,/g' manifest.json
+fi
+
 zip -r ../build.zip *
 cd ..
 rm -rf build
